+++ lib/generators/datatable/templates/ajax_datatables_rails_config.rb
@@ -6,7 +6,4 @@ AjaxDatatablesRails.configure do |config|
 
   # Or you can use your rails environment adapter if you want a generic dev and production
   # config.db_adapter = Rails.configuration.database_configuration[Rails.env]['adapter'].to_sym
-
-  # available options for orm are: :active_record, :mongoid
-  # config.orm = :active_record
 end